+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Next Release
+## 0.5.0 (2017-10-02)
 
 - **[BC]** Remove `Config` in favour of "functional options" in the `options` package
 - **[BC]** Session attributes are now namespaced